Finish renaming publish operation files
